fizz_buzz revisited and moved not really happy with the final result, but interesting for me
